mapping extensions moved to api host
