+++ InnoClinic.sln
@@ -3,29 +3,36 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Appointments.ApiHost", "Innowise.Appointments.ApiHost\Innowise.Appointments.ApiHost.csproj", "{5FE4B873-55C3-4C79-B866-037CE291616E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Appointments.ApiHost", "Innowise.Appointments.ApiHost\Innowise.Appointments.ApiHost.csproj", "{5FE4B873-55C3-4C79-B866-037CE291616E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Appointments.Application", "Innowise.Appointments.Application\Innowise.Appointments.Application.csproj", "{F23D179A-64FF-4167-A92E-8E6D94939A9A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Appointments.Application", "Innowise.Appointments.Application\Innowise.Appointments.Application.csproj", "{F23D179A-64FF-4167-A92E-8E6D94939A9A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Appointments.Contracts", "Innowise.Appointments.Contracts\Innowise.Appointments.Contracts.csproj", "{29927723-3812-49D0-8B5D-673B9E17B00E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Appointments.Contracts", "Innowise.Appointments.Contracts\Innowise.Appointments.Contracts.csproj", "{29927723-3812-49D0-8B5D-673B9E17B00E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Appointments.Domain", "Innowise.Appointments.Domain\Innowise.Appointments.Domain.csproj", "{4D612D11-E708-40A5-8CD5-79EBD0C3757F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Appointments.Domain", "Innowise.Appointments.Domain\Innowise.Appointments.Domain.csproj", "{4D612D11-E708-40A5-8CD5-79EBD0C3757F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Appointments.Infrastructure", "Innowise.Appointments.Infrastructure\Innowise.Appointments.Infrastructure.csproj", "{41BA7D6F-88FB-462F-BBAA-AE44E205C795}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Appointments.Infrastructure", "Innowise.Appointments.Infrastructure\Innowise.Appointments.Infrastructure.csproj", "{41BA7D6F-88FB-462F-BBAA-AE44E205C795}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Appointments.RmqHost", "Innowise.Appointments.RmqHost\Innowise.Appointments.RmqHost.csproj", "{A21A03DD-3463-4B72-A103-E3269892545D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Appointments.RmqHost", "Innowise.Appointments.RmqHost\Innowise.Appointments.RmqHost.csproj", "{A21A03DD-3463-4B72-A103-E3269892545D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Appointments.WorkerHost", "Innowise.Appointments.WorkerHost\Innowise.Appointments.WorkerHost.csproj", "{F275DF7F-C083-4731-82B1-93CD0520C662}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Appointments.WorkerHost", "Innowise.Appointments.WorkerHost\Innowise.Appointments.WorkerHost.csproj", "{F275DF7F-C083-4731-82B1-93CD0520C662}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Innowise.Common", "Innowise.Common\Innowise.Common.csproj", "{A48CDEC8-F552-4B45-A445-3FF9B11B77F7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Common", "Innowise.Common\Innowise.Common.csproj", "{A48CDEC8-F552-4B45-A445-3FF9B11B77F7}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Offices.ApiHost", "Offices.API\Innowise.Offices.ApiHost.csproj", "{891485AB-4EB8-401E-97FB-786F095A3070}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Offices.Application", "Offices.Application\Innowise.Offices.Application.csproj", "{0F0A3188-998F-4710-8DA3-A163D647B0BC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Offices.Domain", "Offices.Domain\Innowise.Offices.Domain.csproj", "{36212E6B-1E03-4E17-92E1-996AB1BDAA6A}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Offices.Infrastructure", "Offices.Infrastructure\Innowise.Offices.Infrastructure.csproj", "{E5B3B0B4-891A-4498-B125-94FDCA0DD4AC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Innowise.Offices.Contracts", "Innowise.Offices.Contracts\Innowise.Offices.Contracts.csproj", "{193111D1-7131-4685-942A-03AD1A40E9CD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{5FE4B873-55C3-4C79-B866-037CE291616E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -60,5 +67,31 @@ Global
 		{A48CDEC8-F552-4B45-A445-3FF9B11B77F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A48CDEC8-F552-4B45-A445-3FF9B11B77F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A48CDEC8-F552-4B45-A445-3FF9B11B77F7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{891485AB-4EB8-401E-97FB-786F095A3070}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{891485AB-4EB8-401E-97FB-786F095A3070}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{891485AB-4EB8-401E-97FB-786F095A3070}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{891485AB-4EB8-401E-97FB-786F095A3070}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0F0A3188-998F-4710-8DA3-A163D647B0BC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0F0A3188-998F-4710-8DA3-A163D647B0BC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0F0A3188-998F-4710-8DA3-A163D647B0BC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0F0A3188-998F-4710-8DA3-A163D647B0BC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36212E6B-1E03-4E17-92E1-996AB1BDAA6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36212E6B-1E03-4E17-92E1-996AB1BDAA6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36212E6B-1E03-4E17-92E1-996AB1BDAA6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36212E6B-1E03-4E17-92E1-996AB1BDAA6A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E5B3B0B4-891A-4498-B125-94FDCA0DD4AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E5B3B0B4-891A-4498-B125-94FDCA0DD4AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E5B3B0B4-891A-4498-B125-94FDCA0DD4AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E5B3B0B4-891A-4498-B125-94FDCA0DD4AC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{193111D1-7131-4685-942A-03AD1A40E9CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{193111D1-7131-4685-942A-03AD1A40E9CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{193111D1-7131-4685-942A-03AD1A40E9CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{193111D1-7131-4685-942A-03AD1A40E9CD}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {54859DC5-1E7A-43FF-A3D8-E608C7C9788D}
 	EndGlobalSection
 EndGlobal
