+++ Innowise.Common/Services/DependencyInjection.cs
@@ -1,13 +1,10 @@
 using System.Net;
 using System.Net.Mail;
 using Innowise.Common.Extensions;
-using Innowise.Common.Services.Authentication;
 using Innowise.Common.Services.Email;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using Microsoft.IdentityModel.Tokens;
 
 namespace Innowise.Common.Services;
